Avoid using getResourceAsStream if possible
http://blog.danlew.net/2013/08/20/joda_time_s_memory_issue_in_android/

Causes a quite substantial memory leak, so do our best to get the zip file
path and open it ourselves when parsing the configuration file. Will
fallback to classloader method if we can't find the zip path, but I haven't
seen that happen on any Android version or device so it should be pretty
foolproof.